+++ CHANGELOG.md
@@ -51,12 +51,17 @@ See https://github.com/dangoslen/changelog-enforcer.
 - [#2388](https://github.com/NibiruChain/nibiru/pull/2388) - chore: erc20 token registry new tokens: cbBTC, uBTC
 - [#2395](https://github.com/NibiruChain/nibiru/pull/2395) - chore: erc20 token registry new token: ynETHx
 - [#2391](https://github.com/NibiruChain/nibiru/pull/2391) - feat(proto): impl script for gRPC Gateway REST doc generation
+- [#2394](https://github.com/NibiruChain/nibiru/pull/2394) - feat(proto): REST API doc generation for bank, auth, and txs
 - [#2397](https://github.com/NibiruChain/nibiru/pull/2397) - feat(epic-evm): rearchitecture for StateDB safety, fix for consensus failures, performance improvements, consistent simulations, and nonce resolution for pending txs in the mempool
 - [#2405](https://github.com/NibiruChain/nibiru/pull/2405) - chore: additional coin logos which could be used externally
 - [#2406](https://github.com/NibiruChain/nibiru/pull/2406) - chore: added monad logo svg
 - [#2407](https://github.com/NibiruChain/nibiru/pull/2407) - feat(sudo-ante): implement zero gas actors for invoking whitelisted contract
 - [#2409](https://github.com/NibiruChain/nibiru/pull/2409) - fix(evm-trace-block): handle native tracer errors JSON-RPC errors for "debug_traceBlockByNumber". Fixes [Nibiru#2400 bug](https://github.com/NibiruChain/nibiru/issues/2400)
-
+- [#2410](https://github.com/NibiruChain/nibiru/pull/2410) -
+feat(evm/grpc-query): Update the "/eth.evm.v1.Query/Balance" query to work with
+"0x" Ethereum hex and "nibi"-prefixed Bech32 address formats. Return no Eth
+address when Wasm contract addresses are queried because 32-byte address space
+(wasm contract length) is not uniquely identified in 20-byte space (EOA length).
 
 ## [v2.7.0](https://github.com/NibiruChain/nibiru/releases/tag/v2.7.0) - 2025-09-15